Update footer.html
I noticed that on this footer you still have links to the stickermule marketplace where the phonegap stickers used to reside.  Please note that stickermule killed their marketplace some time ago. 

We (stickeryou.com) actually already have a phonegap landing page on our marketplace:
http://www.stickeryou.com/products/phonegap/802

As such I'm proposing to change the current: 
<li><a href="https://www.stickermule.com/user/1070635098/stickers">Stickers</a></li>

to:
<li><a href="https://www.stickeryou.com/products/phonegap/802">Stickers</a></li>

On this page we feature stickers, magnets, iron-ons, and temporary tattoos. 

As such I think it would make total sense to change your footer link (which currently points no-where) to our phonegap landing page.   Please review and accept if you feel appropriate.  At this point in time if someone clicks on the current link they will not be able to get anything phonegap related so I think this needs to be fixed.